Moved around the logic a bit.
Moved the transform of the ray to game scene class from physics.
Also implemented a function in physics to take a ray to test it against all bodies.
Created functions in the graphics class to get the view and projection matrices which were needed by the ray transform.
Not working currently because something breaks in the physics, will look into it later.