Reworked some of the code.
Need to fix so that the intersection checks against all bodies and report the closest one instead of reporting the first it hits in the list which it does now.

Changed the ray transform from being manually calculated to using XMVector3Unproject() to remove potential mathematical errors.

Rewrote the Ray/Sphere intersect algorithm to follow the one int the Rel-time Collision book instead of the Real Time Rendering book since i wrote the latter from memory and wasn't sure if it was correct.

Added some logger messages to notify witch body is hit.